% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationIndex.R
\name{CorIndex.all.plusP}
\alias{CorIndex.all.plusP}
\title{Calculated both the non and group-equalised Phi Coefficient of correlation for all habitats
and the p-values.}
\usage{
CorIndex.all.plusP(InDataframe, speciesbinary, weighted, group,
  numberIteration = 1000, SquareID = NULL)
}
\arguments{
\item{InDataframe}{dataframe - containing the data on species presence absence,
habitat and weighting. It also optionally could contain location ID's.}

\item{speciesbinary}{string - the name of the binary field within InDataframe
that contains binary presence absence for a species at the location.}

\item{weighted}{string - name of habitat weighting field. For example, proportion
of habitat, but any weighting with all habitats at a location adding to 1.}

\item{group}{string - ID of habitats or groups of interest.}

\item{numberIteration}{Integer - optional - the number of permutation, set as 1000 by default.}

\item{SquareID}{string - optional if non-weighted, but MUST be used if a weighting
is used. ID of each location. Useful later for permutating the locations to
calculate p-values.}
}
\value{
a custom class "CorIndexVar" containing all values which are the
    same for all habitats and don't need to be recalculated for every
    habitat. Including Nk, nk, the habitats "GroupDF", N, K, Ngp, nkoverNk
    ng & n. See \cite{De Cáceres and Legendre (2009)}. To this is added another custom
    class "Output" as $`result` which gives the habitat ID, R (non group-equalised phi)
    pR (the p-value for R), Rg (group-equalised phi) and pRg (the p-value for Rg). This
    "Output" class is the main result, the CorIndexVar is included as metadata of the
    input values.
}
\description{
This function calculates the non and group-equalised versions of the Phi coefficient of
correlation \cite{Tichy and Chytry (2006), De Cáceres and Legendre (2009)}.
}
\details{
It calculate p-values by permutation and therefore calculates take a bit of time.
This is the main way of useing the PhiCor package and give full outputs.
}
\examples{
    #Using the included dataframe Species1
    Species1_AllPhiPvalues = CorIndex.all.plusP(InDataframe = Species1, speciesbinary = "Species1", weighted = 'Proportion', group = 'HabId', SquareID = 'LocationID')
    names(Species1_AllPhiPvalues)
    Species1_AllPhiPvalues$`result`
}
\references{
Tichy, L., & Chytry, M. (2006). Statistical determination of diagnostic species for site groups of unequal size. Journal of Vegetation Science, 17(6), 809–818.

De Cáceres, M., & Legendre, P. (2009). Associations between species and groups of sites: indices and statistical inference. Ecology, 90(12), 3566–3574.

Future publication...
}
\seealso{
\link[PhiCor]{PhiCor}
}
\author{
Jordan Chetcuti
}
