% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationIndex.R
\name{CorIndex.default}
\alias{CorIndex.default}
\title{The default method of \link[PhiCor]{CorIndex}.}
\description{
This should not be used independently. See \link[PhiCor]{CorIndex} for usage.
}
