% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationIndex.R
\name{CorIndex}
\alias{CorIndex}
\title{CorIndex is basic initiation of variables that are the same for
all habitats in an analysis of multiple habitats in \link[PhiCor]{PhiCor}.}
\usage{
CorIndex(InDataframe, speciesbinary, weighted, group, SquareID = NULL)
}
\arguments{
\item{InDataframe}{dataframe - containing the data on species presence absence,
habitat and weighting. It also optionally could contain location ID's.}

\item{speciesbinary}{string - the name of the binary field within InDataframe
that contains binary presence absence for a species at the location.}

\item{weighted}{string - name of habitat weighting field. For example, proportion
of habitat, but any weighting with all habitats at a location adding to 1.}

\item{group}{string - ID of habitats or groups of interest.}

\item{SquareID}{string - optional ID of each location. Useful later for
permutating the locations to calculate p-values.}
}
\value{
a custom class "CorIndexVar" containing all values which are the
    same for all habitats and don't need to be recalculated for every
    habitat. Including Nk, nk, the habitats "GroupDF", N, K, Ngp, nkoverNk
    ng & n. See \cite{De Cáceres and Legendre (2009)}
    This is reused to give all the results later.
}
\description{
This function is usually called by other function, but could be run
seperatly to give many of the variables which go into the equations
to calculate the phi coefficent. This is useful if manually checking
the calculation.
}
\examples{
    #Using the included dataframe Species1
    Inputs_species1 = CorIndex(InDataframe = Species1, speciesbinary = 'Species1', weighted = 'Proportion', group = 'HabId', 'LocationID')
    Inputs_species1$nkoverNk

    #Using the included dataframe Species2
    #If you wanted the analysis to be non-weighted.
    Species2Unweight = Species2[which(Species2$Proportion==1),]
    #This analysis is then the unweighted version of the analysis. SquareID is not needed.
    Inputs_species2 =  CorIndex(InDataframe = Species2Unweight, speciesbinary = 'Species2', weighted = 'Proportion', group = 'HabId')
    names(Inputs_species2)

}
\references{
De Cáceres, M., & Legendre, P. (2009). Associations between species and groups of sites: indices and statistical inference. Ecology, 90(12), 3566–3574.

Future publication...
}
\author{
Jordan Chetcuti
}
