% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationIndex.R
\name{CorIndex.TargetVar}
\alias{CorIndex.TargetVar}
\title{CorIndex.TargetVar is creates the variables that are particular to a
habitat that were not created in the \link[PhiCor]{CorIndex}}
\usage{
CorIndex.TargetVar(CorIndexVarInput, targetgroup)
}
\arguments{
\item{CorIndexVarInput}{is the custom class "CorIndexVar" output from \link[PhiCor]{CorIndex}}

\item{targetgroup}{Integer - ID of the habitat of interest.}
}
\value{
a custom class "CorIndexVar" containing all values which are the
    same for all habitats and don't need to be recalculated for every
    habitat and the ones for the specific target habitat. Including those from
    CorIndex Nk, nk, the habitats "GroupDF", N, K, Ngp, nkoverNk ng & n and
    aditionally np, Np & ngp  See \cite{De Cáceres and Legendre (2009)}
    This is reused to give all the results later.
}
\description{
This function is usually called by other function, but could be run
seperatly to give many of the variables which go into the equations
to calculate the phi coefficent. This is useful if manually checking
the calculation.
}
\examples{
    #Builing on the examples from \\link[PhiCor]{CorIndex}
    #Using the included dataframe Species1
    Inputs_species1 = CorIndex(InDataframe = Species1, speciesbinary = 'Species1', weighted = 'Proportion', group = 'HabId', 'LocationID')
    Inputs_species1$nkoverNk
    #Using CorIndex.TargetVar, the input is the output of CorIndex. In this case we are interested in
    #habitat 1 (using the ID for the habitat)
    Inputs_species1_hab1 = CorIndex.TargetVar(Inputs_species1, 1)


    #Using the included dataframe Species2
    #If you wanted the analysis to be non-weighted.
    Species2Unweight = Species2[which(Species2$Proportion==1),]
    #This analysis is then the unweighted version of the analysis. SquareID is not needed.
    Inputs_species2 =  CorIndex(InDataframe = Species2Unweight, speciesbinary = 'Species2', weighted = 'Proportion', group = 'HabId')
    names(Inputs_species2)

    Looking at habitat 4
    Inputs_species2_hab4 = CorIndex.TargetVar(Inputs_species2, 4)


}
\references{
De Cáceres, M., & Legendre, P. (2009). Associations between species and groups of sites: indices and statistical inference. Ecology, 90(12), 3566–3574.

Chetcuti, J., Kunin, W. E. & Bullock, J. M. A weighting method to improve habitat association analysis: tested on British carabids. Ecography (Cop.). 42, 1395–1404 (2019).
}
\seealso{
\link[PhiCor]{CorIndex}
}
\author{
Jordan Chetcuti
}
