% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationIndex.R
\name{FilePathInTree}
\alias{FilePathInTree}
\title{Looks for a filename in the current folder and then expands up and along the file tree
looking for the filename.}
\usage{
FilePathInTree(StringFileName)
}
\arguments{
\item{StringFileName}{string filename}
}
\value{
Filepath including the input StringFileName for all files found in the vicinity.
    If it finds multiple files it returns a list of paths which can be accessed individually
    using [1], [2]...etc.
}
\description{
This isn't really part of calculating the Phi coefficient, but is useful in finding a
file which has been reorganised within a project. Imperfect as will return multiple
paths if the filename is used twice locally.
}
\examples{
    #Will return as "character" all R related files in the directory you are currently working on
    PathRfiles = FilePathInTree(".R")
    #If there is only one file PathRfiles would return it, but with multiple PathRfiles[1] would return the first or only.
    PathRfiles[1]

}
