% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationIndex.R
\name{Indicator.all.plusP}
\alias{Indicator.all.plusP}
\title{Calculated both the non and group-equalised IndVal for all habitats
and the p-values.}
\usage{
Indicator.all.plusP(InDataframe, speciesbinary, weighted, group,
  numberIteration = 1000, SquareID = NULL, toroidal = NULL)
}
\arguments{
\item{InDataframe}{dataframe - containing the data on species presence absence,
habitat and weighting. It also optionally could contain location ID's.}

\item{speciesbinary}{string - the name of the binary field within InDataframe
that contains binary presence absence for a species at the location.}

\item{weighted}{string - name of habitat weighting field. For example, proportion
of habitat, but any weighting with all habitats at a location adding to 1.}

\item{group}{string - ID of habitats or groups of interest.}

\item{numberIteration}{Integer - optional - the number of permutation, set as 1000 by default.}

\item{SquareID}{string - optional if non-weighted, but MUST be used if a weighting
is used. ID of each location. Useful later for permutating the locations to
calculate p-values.}

\item{toroidal}{Boolean - If you suspect that your data is spatially auto-correlated you should use 
a toroidal permutation.}
}
\value{
a custom class "CorIndexVar" containing all values which are the
    same for all habitats and don't need to be recalculated for every
    habitat. Including Nk, nk, the habitats "GroupDF", N, K, Ngp, nkoverNk
    ng & n. See \cite{De Cáceres and Legendre (2009)}. To this is added another custom
    class "Output" as $`result` which gives the habitat ID, I (non group-equalised IndVal)
    pI (the p-value for I), Ig (group-equalised IndVal) and Ig (the p-value for Ig). This
    "Output" class is the main result, the CorIndexVar is included as metadata of the
    input values.
}
\description{
This function calculates the non and group-equalised versions of an habitat indicator value (IndVal)
correlation \cite{De Cáceres and Legendre (2009)}.
}
\details{
It calculate p-values by permutation and therefore calculates take a bit of time.
}
\examples{
    #Using the included dataframe Species1
    Species1_AllIndValvalues = Indicator.all.plusP(InDataframe = Species1, speciesbinary = "Species1", weighted = 'Proportion', group = 'HabId', SquareID = 'LocationID')
    names(Species1_AllIndValvalues)
    Species1_AllIndValvalues$`result`
}
\references{
De Cáceres, M., & Legendre, P. (2009). Associations between species and groups of sites: indices and statistical inference. Ecology, 90(12), 3566–3574.

Chetcuti, J., Kunin, W. E. & Bullock, J. M. A weighting method to improve habitat association analysis: tested on British carabids. Ecography (Cop.). 42, 1395–1404 (2019).
}
\seealso{
\link[PhiCor]{PhiCor}
}
\author{
Jordan Chetcuti
}
