% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationIndex.R
\name{Indicator.groupEqual}
\alias{Indicator.groupEqual}
\title{Calculated the group-equalised species indicator value (IndVal).}
\usage{
Indicator.groupEqual(CorIndexTargetVarInput)
}
\arguments{
\item{CorIndexTargetVarInput}{a custom class "CorIndexVar" containing all values from
from both \link[PhiCor]{CorIndex} and \link[PhiCor]{CorIndex.TargetVar}}
}
\value{
IndVal - a floating point number between 0 and 1.
}
\description{
This function calculates the group-equalised version of the species indicator value (IndVal)
\cite{De Cáceres and Legendre (2009)}. This is
useful alone to calculate the group-equalised version for a single habitat of interest
among multiple other habitats.
}
\examples{
    #Builing on the examples from \link[PhiCor]{CorIndex.TargetVar}
    #Using the included dataframe Species1
    Inputs_species1 = CorIndex(InDataframe = Species1, speciesbinary = 'Species1', weighted = 'Proportion', group = 'HabId', 'LocationID')
    Inputs_species1$nkoverNk
    #Using CorIndex.TargetVar, the input is the output of Indicator. In this case we are interested in
    #habitat 1 (using the ID for the habitat)
    Inputs_species1_hab1 = CorIndex.TargetVar(Inputs_species1, 1)

    species1_hab1_GE_IndVal = Indicator.groupEqual(Inputs_species1_hab1)



    #Using the included dataframe Species2
    #If you wanted the analysis to be non-weighted.
    Species2Unweight = Species2[which(Species2$Proportion==1),]
    #This analysis is then the unweighted version of the analysis. SquareID is not needed.
    Inputs_species2 =  CorIndex(InDataframe = Species2Unweight, speciesbinary = 'Species2', weighted = 'Proportion', group = 'HabId')
    names(Inputs_species2)

    #Looking at habitat 4
    Inputs_species2_hab4 = Indicator.TargetVar(Inputs_species2, 4)

    species2_hab4_GE_IndVal = Indicator.groupEqual(Inputs_species2_hab4)


}
\references{
De Cáceres, M., & Legendre, P. (2009). Associations between species and groups of sites: indices and statistical inference. Ecology, 90(12), 3566–3574.

Chetcuti, J., Kunin, W. E. & Bullock, J. M. A weighting method to improve habitat association analysis: tested on British carabids. Ecography (Cop.). 42, 1395–1404 (2019).
}
\seealso{
\link[PhiCor]{PhiCor}
}
\author{
Jordan Chetcuti
}
