% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationIndex.R
\name{Indicator.xPerm}
\alias{Indicator.xPerm}
\title{Permutates and calculates for each permutation both the non and group-equalised
species indicator value (IndVal).}
\usage{
Indicator.xPerm(
  InDataframe,
  FieldToPerm,
  weighted,
  groupsField,
  targetGroup,
  x = 1000,
  SquareID = NULL,
  toroidal = FALSE
)
}
\arguments{
\item{InDataframe}{dataframe - containing the data on species presence absence,
habitat and weighting. It also optionally could contain location ID's.}

\item{FieldToPerm}{string - the name of the binary field within InDataframe
that contains binary presence absence for a species at the location.}

\item{weighted}{string - name of habitat weighting field. For example, proportion
of habitat, but any weighting with all habitats at a location adding to 1.}

\item{groupsField}{string - ID of habitats or groups of interest.}

\item{x}{Integer - optional - the number of permutation, set as 1000 by default.}

\item{SquareID}{string - optional if non-weighted, but MUST be used if a weighting
is used. ID of each location. Useful later for permutating the locations to
calculate p-values.}

\item{toroidal}{Boolean - If you suspect that your data is spatially auto-correlated you should use 
a toroidal permutation.}

\item{targetgroup}{Integer - ID of the habitat of interest.}
}
\value{
Two data frames, one for the group-equalised $`GroupEqualPermList` and one
    for the non group-equalised $NonGroupEqualPermList. It also returns the number of
    permutations $Permutations.
}
\description{
This function purmutaes and calculates the non and group-equalised versions of species indicator 
value (IndVal) \cite{De Cáceres and Legendre (2009)}.
It is useful as it gives the distribution of IndVal values if the species are distributed
randomly across the locations, the IndVal value for the actual data is compared to this
distribution to calculate the p-value. The distribution is a useful output to visualise
the distribution.
}
\examples{
    #Building on the examples from \link[PhiCor]{Indicator.nonEqual}
    #Using the included dataframe Species1
    Inputs_species1 = CorIndex(InDataframe = Species1, speciesbinary = 'Species1', weighted = 'Proportion', group = 'HabId', 'LocationID')
    Inputs_species1$nkoverNk
    #Using CorIndex.TargetVar, the input is the output of Indicator. In this case we are interested in
    #habitat 1 (using the ID for the habitat)
    Inputs_species1_hab1 = CorIndex.TargetVar(Inputs_species1, 1)

    species1_hab1_Indval = Indicator.nonEqual(Inputs_species1_hab1)

    #histogram showing the output of Indicator.xPerm for habitat 1 in comparison to the IndVal value from actual data.
    hist(Hab1_PermutatedPhi$NonGroupEqualPermList, main = "Non group-equalised distribution", xlab = "IndVal values", breaks = 100)
    abline(v=species1_hab1_IndVal, col="red")
    text((species1_hab1_IndVal + 0.006), 15, "Calulated IndVal value", col = "red", srt=90)



}
\seealso{
\link[PhiCor]{PhiCor}
}
\author{
Jordan Chetcuti
}
