% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationIndex.R
\name{lookup}
\alias{lookup}
\title{Lookup function for looking up values from one dataframe for another}
\usage{
lookup(MasterDataField, AdditionField, AdditionInputField)
}
\arguments{
\item{MasterDataField}{vector - a field with common values to those in another dataset you want to
lookup data in.}

\item{AdditionField}{vector - a field with common values to MasterDataField}

\item{AdditonInputField}{vector - the data you are importing to the Master dataframe where the
AdditionField matches the MasterDataField}
}
\value{
vector - will reurn the vector corrosponding to the MasterDataField that
   are found in the lookup field AdditonInputField
}
\description{
Works very similarly to vlookup in excel.
The output is a field in the dataset you are wanting to copy information into.
The syntax is something like:
MyDataframe$CopiedField = lookup(MyDataframe$joinField, OtherDataframe$joinField, OutherDataframe$DataToCopy)
}
\examples{
   #Lookup DF
   HabID = c(1, 2, 3, 4, 5, 6)
   Habitat = c("BLeave", "AGrass", "CGrass", "Water", "Bog", "Heath")
   Desc = c("Broadleaved and mixed woodland", "Acid grassland","Calcareous grassland", "Inland freshwater", "Sphagnum bog", "Caluna heathland")
   OpenClosed = c("Closed", "Open", "Open", "Open","Open","Open")
   DomSpecies = c("Oak", "Bent grass", "Festuca ovina", "NA", "Sphagnum Spp.", "Caluna Spp.")
   lookupdf = data.frame(HabID, Habitat, Desc, OpenClosed, DomSpecies)

   #Feild sites dataframe.
   SiteID = c(1,2,3,4,5,6,7,8,9,10)
   SiteHabitat = c("BLeave", "AGrass", "BLeave", "AGrass", "Heath", "Bog", "Water", "BLeave", "Water", "Heath")
   SitesDF = data.frame(SiteID, SiteHabitat)

   #Looking up the descriptions of the habitats.
   SitesDF$HabDesc = lookup(SitesDF$SiteHabitat, lookupdf$Habitat, lookupdf$Desc)

   #You could lookup other information from the lookup table.
}
