% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationIndex.R
\docType{data}
\name{Species2}
\alias{Species2}
\title{Species 2 data of presence absence and weighting for LCM2015 land cover}
\format{A data frame with  1,005 rows and 5 variables:}
\usage{
The data is useful in examples of how to use the functions.
\describe{
  \item{LocationID}{A anonymised location}
  \item{bhab}{LCM2015 land cover description}
  \item{Proportion}{The proportion of a location that is each habitat}
  \item{HabId}{An integer ID for each habitat in LCM2015}
  \item{Species2}{Binary presence absence for "species2"}
}
}
\description{
The data is anonymised data created using LCM2015 \cite{Rowland et al., (2017)} and \cite{NBN Atlas website (2018)}
}
\references{
Rowland, C. S., Morton, R. D., Carrasco, L., McShane, G., O’Neil, A. W., & Wood, C. M. (2017). Land Cover Map 2015 (vector, GB). NERC Environmental Information Data Centre. NERC Environmental Information Data Centre. doi:/10.5285/6c6c9203-7333-4d96-88ab-78925e7a4e73

NBN Atlas website. (2018). at https://nbnatlas.org. Accessed 13 April 2018.
}
\keyword{datasets}
