% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationIndex.R
\docType{package}
\name{PhiCor-package}
\alias{PhiCor}
\alias{PhiCor-package}
\title{Package for calculation the Weighted Phi Coefficient of association}
\description{
This package can be used to calculate the weighted phi coeficient of
association using a weighting for habitats within a location, or
can calculate the unweighted original version from
\cite{De Cáceres and Legendre (2009)} by using a weighting of one.
The package also allows the group-equalised version to be run.
}
\details{
The Phi Coefficent of association gives the habitat preference, or
site fidelity of habitat sites.


    The main function in PhiCor is \link[PhiCor]{CorIndex.all.plusP}.
    This will calculate all phi corelation of association values and
    purmutate to give p-values.

    The other Phi coefficent functions are useful in testing or calculating only part
    of the full outputs from \link[PhiCor]{CorIndex.all.plusP}:
    \itemize{
        \item \link[PhiCor]{CorIndex} calculates the inputs that don't change between habitats.
        \item \link[PhiCor]{CorIndex.TargetVar} calculates the additional input specific to a habitat.
        \item \link[PhiCor]{CorIndex.groupEqual} calculates the group-equalised phi for a habitat.
        \item \link[PhiCor]{CorIndex.nonEqual} calculates the non group-equalised phi for a habitat.
        \item \link[PhiCor]{CorIndex.xPerm} returns the null phi distribution for a habitat.
        \item \link[PhiCor]{CorIndex.all} returns the non and group-equalised phi values for all habitats
            without p-values and is therefore quicker.
    }
    Weighted habitat indicator values (IndVal) can also be calculated, but these have not been tested in
    the way that the phi coefficent has been. 
    The main function for calculating IndVal is \link[PhiCor]{Indicator.all.plusP}
    These functions are useful in testing or calculating only part
    of the full outputs from \link[PhiCor]{Indicator.all.plusP}:
    \itemize{
        \item \link[PhiCor]{Indicator.nonEqual} Calculated the non group-equalised habitat indicator value (IndVal).
        \item \link[PhiCor]{Indicator.groupEqual} Calculated the group-equalised species indicator value (IndVal).
        \item \link[PhiCor]{Indicator.xPerm} Permutates and calculates for each permutation both the non and 
             group-equalised species indicator value (IndVal).
        \item 
    } 

    The functions \link[PhiCor]{lookup} and \link[PhiCor]{FilePathInTree}
    might be used elsewhere.
}
\references{
De Cáceres, M., & Legendre, P. (2009). Associations between species and groups of sites: indices and statistical inference. Ecology, 90(12), 3566–3574.

Chetcuti, J., Kunin, W. E. & Bullock, J. M. A weighting method to improve habitat association analysis: tested on British carabids. Ecography (Cop.). 42, 1395–1404 (2019).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://doi.org/10.5281/zenodo.1407053}
}

}
\author{
Jordan Chetcuti
}
\keyword{internal}
